# backend/routes/qdrive.mjs
import expressPkg from 'express';
const router = expressPkg.Router();

// Upload file to Qdrive
router.post('/upload', async (req, res) => {
  try {
    const { filename, data, userDID, spaceDID } = req.body;
    
    console.log(`[Qdrive] Upload file: ${filename} for user: ${userDID}`);
    
    // In a real implementation:
    // 1. Encrypt file with user's key
    // 2. Upload to IPFS via Storacha
    // 3. Store file metadata in database
    // 4. Update user's file index
    
    const fileId = `file_${Date.now()}`;
    const cid = `Qm${Math.random().toString(36).substring(2, 15)}`;
    
    res.json({
      success: true,
      fileId,
      cid,
      filename,
      uploadedAt: new Date().toISOString()
    });
    
  } catch (error) {
    console.error('[Qdrive] Upload error:', error);
    res.status(500).json({ 
      success: false,
      error: 'Upload failed' 
    });
  }
});

// Get user files
router.get('/files/:userDID', async (req, res) => {
  try {
    const { userDID } = req.params;
    
    console.log(`[Qdrive] Get files for user: ${userDID}`);
    
    // In a real implementation, query database for user's files
    const files = [];
    
    res.json({
      success: true,
      files,
      count: files.length
    });
    
  } catch (error) {
    console.error('[Qdrive] Get files error:', error);
    res.status(500).json({ 
      success: false,
      error: 'Failed to retrieve files' 
    });
  }
});

export default router;