# modules/qdrive/openapi.yaml
openapi: 3.0.3
info:
  title: Qdrive API
  description: Decentralized file storage with IPFS integration and encryption
  version: 1.0.0
  contact:
    name: AnarQ Team
    url: https://github.com/anarq/ecosystem
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3008
    description: Development server
  - url: https://qdrive.anarq.org
    description: Production server
paths:
...
  /files/{cid}:
    get:
      summary: Download file
      operationId: downloadFile
      security:
        - squidAuth: []
      parameters:
        - name: cid
          in: path
          required: true
          schema:
            type: string
          description: IPFS Content ID
        - name: download
          in: query
          schema:
            type: boolean
            default: false
          description: Force download instead of inline display
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete file
      operationId: deleteFile
      security:
        - squidAuth: []
      parameters:
        - name: cid
          in: path
          required: true
          schema:
            type: string
          description: IPFS Content ID
      responses:
        '200':
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    squidAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: sQuid identity token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok]
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        code:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        retryable:
          type: boolean
        suggestedActions:
          type: array
          items:
            type: string